+++ src/features/firstAid/ui/FirstAidResult.tsx
@@ -1,68 +1,28 @@
 'use client';
 
-import { useEffect, useState, useRef } from 'react';
-import { useSymptomStore } from '@/src/features/firstAid/store/useSymptomStore';
-import { postFirstAid } from '@/src/features/firstAid/api/firstAid';
-import { useHydration } from '@/src/shared/hooks/useHydration';
 import { motion, AnimatePresence, Variants } from 'motion/react';
 
 import { AiOutlineFileText, AiOutlineLoading3Quarters } from 'react-icons/ai';
 import { BsCheckCircle, BsExclamationTriangle } from 'react-icons/bs';
 import { HiOutlineChartBar } from 'react-icons/hi';
 import { FaRegLightbulb, FaArrowRight } from 'react-icons/fa';
+import useFirstAidResult from '../hooks/useFirstAidResult';
 
-interface FirstAidResult {
-  content: string;
-  recommendedAction: string;
-  confidence: number;
-  blogLinks: string[];
-}
+const fadeInUp = {
+  hidden: { opacity: 0, y: 20 },
+  visible: (i: number) => ({
+    opacity: 1,
+    y: 0,
+    transition: {
+      delay: i * 0.15,
+      duration: 0.5,
+      ease: 'easeOut',
+    },
+  }),
+} as const satisfies Variants;
 
 export default function FirstAidResult() {
-  const hydrated = useHydration();
-  const { symptomType, symptomDetail } = useSymptomStore();
-  const [result, setResult] = useState<FirstAidResult | string | null>(null);
-  const [loading, setLoading] = useState(true);
-  const contentRef = useRef<HTMLDivElement>(null);
-
-  const fadeInUp = {
-    hidden: { opacity: 0, y: 20 },
-    visible: (i: number) => ({
-      opacity: 1,
-      y: 0,
-      transition: {
-        delay: i * 0.15,
-        duration: 0.5,
-        ease: 'easeOut',
-      },
-    }),
-  } as const satisfies Variants;
-
-  useEffect(() => {
-    if (!hydrated || !symptomType || !symptomDetail) return;
-
-    const postData = async () => {
-      setLoading(true);
-      try {
-        const res = await postFirstAid({ symptomType, symptomDetail });
-
-        setResult(res.result);
-      } catch (e) {
-        setResult('An error occurred while processing your request. Please try again later.');
-        console.error(e);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    postData();
-  }, [hydrated, symptomType, symptomDetail]);
-
-  useEffect(() => {
-    if (result && contentRef.current) {
-      contentRef.current.scrollIntoView({ behavior: 'smooth', block: 'start' });
-    }
-  }, [result]);
+  const { hydrated, symptomType, symptomDetail, result, loading, contentRef } = useFirstAidResult();
 
   if (!hydrated) {
     return (
